+++ site/src/App/ThemeSetting/index.ts
@@ -1,7 +1,3 @@
 export { ThemedExample } from './ThemedExample';
-export {
-  ThemeSettingProvider,
-  ThemeKey,
-  useThemeSettings,
-} from './ThemeSettingContext';
+export { ThemeSettingProvider, useThemeSettings } from './ThemeSettingContext';
 export { ThemeToggle } from './ThemeToggle';
